Remove toolchain registration from rules_cc (#90)
* Remove toolchain registration from rules_cc

This is because Bazel currently uses the built-in cc toolchain configuring mechanism (in `@bazel_tools`), and *not* the one in rules_cc. In order to prevent the local cc toolchain from being discovered twice, this commit removes the automatic toolchain registration from rules_cc. In the rare case where someone wants to specifically use the toolchain from rules_cc, they can call `use_extension` on this extension themselves and add the relevant toolchain in their own MODULE.bazel file.

NB: We're directly updating the current version instead of creating a new version because the BCR is not officially launched yet.

* fix checksum